requirements.txt
transformers
torch
torchaudio
gradio
opencv-python-headless
fer
openai-whisper
numpy
scipy
accelerate